requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
streamlit==1.28.2
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
pytest==8.0.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9 